+++ src/com/dotcms/cluster/business/ClusterFactory.java
@@ -80,14 +80,14 @@ public class ClusterFactory {
 		DotConnect dc = new DotConnect();
 		dc.setSQL("select max(cache_port) as port from server where ip_address = (select s.ip_address from server s where s.server_id = ?)");
 		dc.addParam(serverId);
-		Short maxPort = null;
+		Integer maxPort = null;
 		String freePort = Config.getStringProperty("CACHE_BINDPORT", "7800");
 
 		try {
 			List<Map<String,Object>> results = dc.loadObjectResults();
 			if(!results.isEmpty()) {
-				maxPort = (Short) results.get(0).get("port");
-				freePort = UtilMethods.isSet(maxPort)?Long.toString(maxPort+1):freePort;
+				maxPort = (Integer) results.get(0).get("port");
+				freePort = UtilMethods.isSet(maxPort)?Integer.toString(maxPort+1):freePort;
 			}
 
 		} catch (DotDataException e) {
